update register dan login
