+++ app/src/main/AndroidManifest.xml
@@ -15,14 +15,14 @@
         android:theme="@style/Theme.Organeco"
         tools:targetApi="31">
         <activity
-            android:name=".view.StartedActivity"
+            android:name=".view.start.StartedActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".view.OnboardingActivity"
+            android:name=".view.start.OnboardingActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
@@ -36,13 +36,13 @@
                 android:value="" />
         </activity>
         <activity
-            android:name=".view.LoginActivity"
+            android:name=".view.login.LoginActivity"
             android:exported="false" />
         <activity
-            android:name=".view.RegisterActivity"
+            android:name=".view.register.RegisterActivity"
             android:exported="false" />
         <activity
-            android:name=".view.SplashScreenActivity"
+            android:name=".view.start.SplashScreenActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
