--- app/src/main/java/com/organeco/view/OnboardingActivity.kt
@@ -1,24 +1,0 @@
-package com.organeco.view
-
-import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import com.organeco.R
-import com.organeco.databinding.ActivityOnboardingBinding
-
-class OnboardingActivity : AppCompatActivity() {
-
-    private lateinit var binding: ActivityOnboardingBinding
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityOnboardingBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-        supportActionBar?.hide()
-
-        binding.btnStarted.setOnClickListener {
-            startActivity(Intent(this,RegisterActivity::class.java))
-            finishAffinity()
-        }
-    }
-}
