+++ app/src/main/java/com/organeco/view/start/StartedActivity.kt
@@ -1,0 +1,24 @@
+package com.organeco.view.start
+
+import android.content.Intent
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import com.organeco.databinding.ActivityStartedBinding
+import com.organeco.view.register.RegisterActivity
+
+class StartedActivity : AppCompatActivity() {
+    private lateinit var binding: ActivityStartedBinding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = ActivityStartedBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+
+        supportActionBar?.hide()
+
+        binding.btnStart.setOnClickListener {
+            startActivity(Intent(this, RegisterActivity::class.java))
+            finishAffinity()
+        }
+    }
+}
