+++ app/src/main/res/layout/activity_started.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/green_700"
-    tools:context=".view.StartedActivity">
+    tools:context=".view.start.StartedActivity">
 
     <ImageView
         android:id="@+id/ivPlant1"
